refactor: move code to different files/structure
