+++ src/services/domainService.js
@@ -1,19 +1,15 @@
 /* eslint-disable no-restricted-syntax */
 const axios = require('axios');
-const qs = require('qs');
 const config = require('config');
-const dns = require('dns').promises;
 const nodecmd = require('node-cmd');
 const util = require('util');
 const fs = require('fs').promises;
-const fsSync = require('fs');
-const https = require('https');
 const log = require('../lib/log');
-const serviceHelper = require('./serviceHelper');
 const ipService = require('./ipService');
 const fluxService = require('./fluxService');
 const haproxyTemplate = require('./haproxyTemplate');
 const applicationChecks = require('./applicationChecks');
+const { processApplications, getUnifiedDomains } = require('./domain');
 
 let myIP = null;
 let myFDMnameORip = null;
@@ -25,198 +21,6 @@ const axiosConfig = {
 };
 
 const cmdAsync = util.promisify(nodecmd.run);
-
-let db = null;
-const recordsCollection = config.database.mainDomain.collections.records;
-
-const cloudFlareAxiosConfig = {
-  headers: {
-    'Content-Type': 'application/json',
-    Authorization: `Bearer ${config.cloudflare.apiKey}`,
-  },
-};
-
-// set rejectUnauthorized to false to accept self signed certificates.
-const agent = new https.Agent({
-  rejectUnauthorized: false,
-});
-const pDNSAxiosConfig = {
-  headers: {
-    'X-API-Key': config.pDNS.apiKey,
-  },
-  httpsAgent: agent,
-};
-
-// return array of IPs to which a hostname is pointeed
-async function dnsLookup(hostname) {
-  const result = await dns.lookup(hostname, { all: true }); // eg. [ { address: '65.21.189.1', family: 4 } ]
-  return result;
-}
-
-async function isDomainPointedToThisFDM(hostname, ip = myIP) {
-  try {
-    if (!ip) {
-      return false;
-    }
-    const dnsLookupdRecords = await dnsLookup(hostname);
-    const pointedToMyIp = dnsLookupdRecords.find((record) => record.address === ip);
-    if (pointedToMyIp) {
-      return true;
-    }
-    return false;
-  } catch (error) {
-    log.warn(error);
-    return false;
-  }
-}
-
-// Lists DNS records for given input, will return all if no input provided
-async function listDNSRecords(name, content, type = 'all', page = 1, perPage = 100, records = []) {
-  // https://api.cloudflare.com/#dns-records-for-a-zone-list-dns-records
-  if (config.cloudflare.enabled) {
-    const query = {
-      name,
-      content,
-      page,
-      per_page: perPage,
-    };
-    if (type !== 'all') {
-      query.type = type;
-    }
-    const queryString = qs.stringify(query);
-    const url = `${config.cloudflare.endpoint}zones/${config.cloudflare.zone}/dns_records?${queryString}`;
-    const response = await axios.get(url, cloudFlareAxiosConfig);
-    if (response.data.result_info.total_pages > page) {
-      const recs = records.concat(response.data.result);
-      return listDNSRecords(name, content, type, page + 1, perPage, recs);
-    }
-    const r = records.concat(response.data.result);
-    return r;
-  } if (config.pDNS.enabled) {
-    let adjustedName = name;
-    if (!name) {
-      adjustedName = '*';
-    }
-    const url = `${config.pDNS.endpoint}search-data?q=${adjustedName}&object_type=record`;
-    const response = await axios.get(url, pDNSAxiosConfig);
-    let filteredData = response.data;
-    if (content) {
-      filteredData = filteredData.filter((data) => data.content === content);
-    }
-    if (type !== 'all') {
-      filteredData = filteredData.filter((data) => data.type === type);
-    }
-    return filteredData;
-  }
-  throw new Error('No DNS provider is enable!');
-}
-
-// Deletes DNS record for given id (for cloudflare)
-async function deleteDNSRecordCloudflare(record) {
-  if (!record) {
-    throw new Error('No DNS record specified');
-  }
-  const { id } = record;
-  if (!id) {
-    throw new Error('No DNS ID record specified');
-  }
-  if (!record.name.endsWith(`${config.appSubDomain}.${config.mainDomain}`)) {
-    throw new Error('Invalid DNS record to delete specified');
-  }
-  // https://api.cloudflare.com/#dns-records-for-a-zone-delete-dns-record
-  const url = `${config.cloudflare.endpoint}zones/${config.cloudflare.zone}/dns_records/${id}`;
-  const response = await axios.delete(url, cloudFlareAxiosConfig);
-  return response.data;
-}
-
-// Deletes DNS records matching given parameters (for pDNS)
-async function deleteDNSRecordPDNS(name, content, type = 'A', ttl = 60) {
-  if (!name.endsWith(`${config.appSubDomain}.${config.mainDomain}`)) {
-    throw new Error('Invalid DNS record to delete specified');
-  }
-  if (config.pDNS.enabled) {
-    const data = {
-      rrsets: [{
-        name,
-        type,
-        ttl,
-        changetype: 'DELETE',
-        records: [{ content, disabled: false }],
-      }],
-    };
-    const url = `${config.pDNS.endpoint}zones/${config.pDNS.zone}`;
-    const response = await axios.patch(url, data, pDNSAxiosConfig);
-    return response.data;
-  }
-  throw new Error('No DNS provider is enable!');
-}
-
-// Creates new DNS record
-async function createDNSRecord(name, content, type = config.domainAppType, ttl = 60, skipDNSRecordCreation = config.skipDNSRecordCreation) {
-  if (skipDNSRecordCreation) {
-    return true;
-  }
-  if (!name.endsWith(`${config.appSubDomain}.${config.mainDomain}`)) {
-    throw new Error('Invalid DNS record to create specified');
-  }
-  if (config.cloudflare.enabled) {
-    // https://api.cloudflare.com/#dns-records-for-a-zone-create-dns-record
-    const data = {
-      type,
-      name,
-      content,
-      ttl,
-    };
-    const url = `${config.cloudflare.endpoint}zones/${config.cloudflare.zone}/dns_records`;
-    const response = await axios.post(url, data, cloudFlareAxiosConfig);
-    return response.data;
-  } if (config.pDNS.enabled) {
-    let adjustedContent = content;
-    if (type === 'CNAME') {
-      adjustedContent = `${content}.`;
-    }
-    const data = {
-      rrsets: [{
-        name: `${name}.`,
-        type,
-        ttl,
-        changetype: 'REPLACE',
-        records: [{ content: adjustedContent, disabled: false }],
-      }],
-    };
-    const url = `${config.pDNS.endpoint}zones/${config.pDNS.zone}`;
-    const response = await axios.patch(url, data, pDNSAxiosConfig);
-    return response.data;
-  }
-  throw new Error('No DNS provider is not enabled!');
-}
-
-async function getAllRecordsDBAPI(req, res) {
-  try {
-    const database = db.db(config.database.mainDomain.database);
-    const q = {};
-    const p = {};
-    const records = await serviceHelper.findInDatabase(database, recordsCollection, q, p);
-    const resMessage = serviceHelper.createDataMessage(records);
-    res.json(resMessage);
-  } catch (error) {
-    log.error(error);
-    const errMessage = serviceHelper.createErrorMessage(error.message, error.name, error.code);
-    res.json(errMessage);
-  }
-}
-
-async function listDNSRecordsAPI(req, res) {
-  try {
-    const records = await listDNSRecords();
-    const resMessage = serviceHelper.createDataMessage(records);
-    res.json(resMessage);
-  } catch (error) {
-    log.error(error);
-    const errMessage = serviceHelper.createErrorMessage(error.message, error.name, error.code);
-    res.json(errMessage);
-  }
-}
 
 // Generates config file for HAProxy
 async function generateAndReplaceMainHaproxyConfig() {
@@ -298,31 +102,6 @@ async function getApplicationLocation(appName) {
     return [];
   }
 }
-// generates domain names for a given app specificatoin
-function getUnifiedDomainsForApp(specifications) {
-  const domains = [];
-  const lowerCaseName = specifications.name.toLowerCase();
-  if (specifications.version <= 3) { // app v1 cannot be spawned and do not exist
-    // adding names for each port with new scheme {appname}_{portnumber}.app2.runonflux.io
-    for (let i = 0; i < specifications.ports.length; i += 1) {
-      const portDomain = `${lowerCaseName}_${specifications.ports[i]}.${config.appSubDomain}.${config.mainDomain}`;
-      domains.push(portDomain);
-    }
-  } else {
-    // composed app
-    for (const component of specifications.compose) {
-      // same for composed apps, adding for each port with new scheme {appname}_{portnumber}.app2.runonflux.io
-      for (let i = 0; i < component.ports.length; i += 1) {
-        const portDomain = `${lowerCaseName}_${component.ports[i]}.${config.appSubDomain}.${config.mainDomain}`;
-        domains.push(portDomain);
-      }
-    }
-  }
-  // finally push general name which is alias to first port
-  const mainDomain = `${lowerCaseName}.${config.appSubDomain}.${config.mainDomain}`;
-  domains.push(mainDomain);
-  return domains;
-}
 
 function getCustomConfigs(specifications) {
   const configs = [];
@@ -393,266 +172,9 @@ function getCustomConfigs(specifications) {
   return configs;
 }
 
-function getCustomDomainsForApp(app) {
-  const domains = [];
-  if (app.version <= 3) {
-    for (let i = 0; i < app.ports.length; i += 1) {
-      const portDomains = app.domains[i].split(',');
-      portDomains.forEach((portDomain) => {
-        if (portDomain && portDomain.includes('.') && portDomain.length >= 3 && !portDomain.toLowerCase().endsWith(`${config.appSubDomain}.${config.mainDomain}`)) {
-          let domain = portDomain.replace('https://', '').replace('http://', '').replace(/[&/\\#,+()$~%'":*?<>{}]/g, ''); // . is allowed
-          if (domain.includes('www.')) {
-            // eslint-disable-next-line prefer-destructuring
-            domain = domain.split('www.')[1];
-          }
-          domains.push(domain.toLowerCase());
-          domains.push(`www.${domain.toLowerCase()}`);
-          domains.push(`test.${domain.toLowerCase()}`);
-        }
-      });
-    }
-  } else {
-    for (const component of app.compose) {
-      for (let i = 0; i < component.ports.length; i += 1) {
-        const portDomains = component.domains[i].split(',');
-        portDomains.forEach((portDomain) => {
-          if (portDomain && portDomain.includes('.') && portDomain.length >= 3 && !portDomain.toLowerCase().endsWith(`${config.appSubDomain}.${config.mainDomain}`)) {
-            let domain = portDomain.replace('https://', '').replace('http://', '').replace(/[&/\\#,+()$~%'":*?<>{}]/g, ''); // . is allowed
-            if (domain.includes('www.')) {
-            // eslint-disable-next-line prefer-destructuring
-              domain = domain.split('www.')[1];
-            }
-            domains.push(domain.toLowerCase());
-            domains.push(`www.${domain.toLowerCase()}`);
-            domains.push(`test.${domain.toLowerCase()}`);
-          }
-        });
-      }
-    }
-  }
-  return domains;
-}
-
-// return true if some domain operation was done
-// return false if no domain operation was done
-async function checkAndAdjustDNSrecordForDomain(domain) {
-  try {
-    const dnsRecords = await listDNSRecords(domain);
-    // delete bad
-    for (const record of dnsRecords) { // async inside
-      let adjustedRecord = record.content;
-      if (adjustedRecord && config.pDNS.enabled) {
-        adjustedRecord = adjustedRecord.slice(0, -1);
-      }
-      if (myFDMnameORip && typeof myFDMnameORip === 'string' && record.content && (adjustedRecord !== myFDMnameORip || record.proxied === true)) {
-        // delete the record
-        if (config.cloudflare.enabled) {
-          // eslint-disable-next-line no-await-in-loop
-          await deleteDNSRecordCloudflare(record); // may throw
-          log.info(`Record ${record.id} of ${record.name} on ${record.content} deleted`);
-        } else if (config.pDNS.enabled) {
-          // eslint-disable-next-line no-await-in-loop
-          await deleteDNSRecordPDNS(record.name, record.content, record.type, record.ttl); // may throw
-          log.info(`Record ${record.name} on ${record.content} deleted`);
-        }
-      }
-    }
-    const correctRecords = dnsRecords.filter((record) => ((record.content === myFDMnameORip || (record.content && record.content.slice(0, -1) === myFDMnameORip)) && (record.proxied === undefined || record.proxied === false)));
-    if (correctRecords.length === 0) {
-      await createDNSRecord(domain, myFDMnameORip, config.domainAppType);
-      return true;
-    }
-    if (correctRecords.length > 1) {
-      // delete all except the first one
-      correctRecords.shift(); // remove first record from records to delete
-      for (const record of correctRecords) { // async inside
-        // delete the record
-        if (config.cloudflare.enabled) {
-          // eslint-disable-next-line no-await-in-loop
-          await deleteDNSRecordCloudflare(record); // may throw
-          log.info(`Duplicate Record ${record.id} of ${record.name} on ${record.content} deleted`);
-        } else if (config.pDNS.enabled) {
-          // eslint-disable-next-line no-await-in-loop
-          await deleteDNSRecordPDNS(record.name, record.content, record.type, record.ttl); // may throw
-          log.info(`Duplicate Record ${record.name} on ${record.content} deleted`);
-        }
-      }
-      return true;
-    }
-    // only one record exists and is correct
-    log.info(`Record for domain ${domain} is set correctly`);
-    return false;
-  } catch (error) {
-    log.error(error);
-    return true;
-  }
-}
-
-async function checkCertificatePresetForDomain(domain) {
-  try {
-    const path = `/etc/ssl/fluxapps/${domain}.pem`;
-    await fs.access(path); // only check if file exists. Does not check permissions
-    const fileSize = fsSync.statSync(path).size;
-    if (fileSize > 128) { // it can be an empty file.
-      return true;
-    }
-    return false;
-  } catch (error) {
-    return false;
-  }
-}
-
-async function obtainDomainCertificate(domain) { // let it throw
-  const cmdToExec = `sudo certbot certonly --standalone -d ${domain} --non-interactive --agree-tos --email ${config.emailDomain} --http-01-port=8787`;
-  const cmdToExecContinue = `sudo cat /etc/letsencrypt/live/${domain}/fullchain.pem /etc/letsencrypt/live/${domain}/privkey.pem | sudo tee /etc/ssl/${config.certFolder}/${domain}.pem`;
-  const response = await cmdAsync(cmdToExec);
-  if (response.includes('Congratulations') || response.includes('Certificate not yet due for renewal')) {
-    await cmdAsync(cmdToExecContinue);
-  }
-}
-
-async function adjustAutoRenewalScriptForDomain(domain) { // let it throw
-  const path = '/opt/update-certs.sh';
-  try {
-    await fs.readFile(path);
-    const autoRenewScript = await fs.readFile(path, { encoding: 'utf-8' });
-    const cert = `bash -c "cat /etc/letsencrypt/live/${domain}/fullchain.pem /etc/letsencrypt/live/${domain}/privkey.pem > /etc/ssl/${config.certFolder}/${domain}.pem"`;
-    if (autoRenewScript.includes(cert)) {
-      return;
-    }
-    // split the contents by new line
-    const lines = autoRenewScript.split(/\r?\n/);
-    lines.splice(6, 0, cert); // push cert to top behind #Concatenate...
-    const file = lines.join('\n');
-    await fs.writeFile(path, file, {
-      mode: 0o755,
-      flag: 'w',
-      encoding: 'utf-8',
-    });
-  } catch (error) {
-    // probably does not exist
-    const beginning = `#!/usr/bin/env bash
-# Renew the certificate
-certbot renew --force-renewal --http-01-port=8787 --preferred-challenges http
-# Concatenate new cert files, with less output (avoiding the use tee and its output to stdout)\n`;
-    const ending = `
-# Reload  HAProxy
-service haproxy reload`;
-    const cert = `bash -c "cat /etc/letsencrypt/live/${domain}/fullchain.pem /etc/letsencrypt/live/${domain}/privkey.pem > /etc/ssl/${config.certFolder}/${domain}.pem"\n`;
-    const file = beginning + cert + ending;
-    await fs.writeFile(path, file, {
-      mode: 0o755,
-      flag: 'w',
-      encoding: 'utf-8',
-    });
-  }
-}
-
 async function createSSLDirectory() {
   const dir = `/etc/ssl/${config.certFolder}`;
   await fs.mkdir(dir, { recursive: true });
-}
-
-async function doCertOperationsForCustomDomains(domains) {
-  try {
-    for (const appDomain of domains) {
-      // no dns check
-      if (config.automateCertificates) {
-        try {
-          // check if we have certificate
-          // eslint-disable-next-line no-await-in-loop
-          const isCertificatePresent = await checkCertificatePresetForDomain(appDomain);
-          if (appDomain.length > 64) {
-            log.warn(`Domain ${appDomain} is too long. Certificate not issued`);
-            return true;
-          }
-          if (!isCertificatePresent) {
-            // eslint-disable-next-line no-await-in-loop
-            const domainIsPointedCorrectly = await isDomainPointedToThisFDM(appDomain);
-            if (!domainIsPointedCorrectly) {
-              throw new Error(`DNS record is not pointed to this FDM for ${appDomain}, cert operations not proceeding`);
-            }
-            // if we dont have certificate, obtain it
-            log.info(`Obtianing certificate for ${appDomain}`);
-            // eslint-disable-next-line no-await-in-loop
-            await obtainDomainCertificate(appDomain);
-          }
-          // eslint-disable-next-line no-await-in-loop
-          const isCertificatePresentB = await checkCertificatePresetForDomain(appDomain);
-          if (isCertificatePresentB) {
-            // check if domain has autorenewal, if not, adjust it
-            // eslint-disable-next-line no-await-in-loop
-            await adjustAutoRenewalScriptForDomain(appDomain);
-          } else {
-            throw new Error(`Certificate not present for ${appDomain}`);
-          }
-        } catch (error) {
-          log.warn(error);
-        }
-      }
-    }
-    return true;
-  } catch (error) {
-    log.error(error);
-    return false;
-  }
-}
-
-async function doDomainCertOperationsForFDMdomains(domains) {
-  try {
-    for (const appDomain of domains) {
-      if (appDomain === 'ethereumnodelight.app.runonflux.io') { // temporarily disable
-        // eslint-disable-next-line no-continue
-        continue;
-      }
-      // check DNS
-      // if DNS was adjusted for this domain, wait a minute
-      // eslint-disable-next-line no-await-in-loop
-      const wasDomainAdjusted = await checkAndAdjustDNSrecordForDomain(appDomain);
-      if (wasDomainAdjusted) {
-        log.info(`Domain ${appDomain} was adjusted on DNS`);
-        // eslint-disable-next-line no-await-in-loop
-        await serviceHelper.timeout(1 * 1000);
-      }
-      if (config.automateCertificatesForFDMdomains) {
-        try {
-          // check if we have certificate
-          // eslint-disable-next-line no-await-in-loop
-          const isCertificatePresent = await checkCertificatePresetForDomain(appDomain);
-          if (appDomain.length > 64) {
-            log.warn(`Domain ${appDomain} is too long. Certificate not issued`);
-            return true;
-          }
-          if (!isCertificatePresent) {
-            // eslint-disable-next-line no-await-in-loop
-            const domainIsPointedCorrectly = await isDomainPointedToThisFDM(appDomain);
-            if (!domainIsPointedCorrectly) {
-              throw new Error(`DNS record is not pointed to this FDM for ${appDomain}, cert operations not proceeding`);
-            }
-            // if we dont have certificate, obtain it
-            log.info(`Obtaining certificate for ${appDomain}`);
-            // eslint-disable-next-line no-await-in-loop
-            await obtainDomainCertificate(appDomain);
-          }
-          // eslint-disable-next-line no-await-in-loop
-          const isCertificatePresentB = await checkCertificatePresetForDomain(appDomain);
-          if (isCertificatePresentB) {
-            // check if domain has autorenewal, if not, adjust it
-            // eslint-disable-next-line no-await-in-loop
-            await adjustAutoRenewalScriptForDomain(appDomain);
-          } else {
-            throw new Error(`Certificate not present for ${appDomain}`);
-          }
-        } catch (error) {
-          log.warn(error);
-        }
-      }
-    }
-    return true;
-  } catch (error) {
-    log.error(error);
-    return false;
-  }
 }
 
 // periodically keeps HAproxy ans certificates updated every 4 minutes
@@ -668,77 +190,9 @@ async function generateAndReplaceMainApplicationHaproxyConfig() {
     // add to renewal script
     // check if certificate exist
     // if all ok, add for creation of domain
-    const appsOK = [];
     await createSSLDirectory();
     log.info('SSL directory checked');
-    for (const appSpecs of applicationSpecifications) {
-      // exclude blacklisted apps
-      if (serviceHelper.matchRule(appSpecs.name, config.blackListedApps)) {
-        // eslint-disable-next-line no-continue
-        continue;
-      }
-
-      log.info(`Adjusting domains and ssl for ${appSpecs.name}`);
-      if (appSpecs.name === 'themok') {
-        for (const component of appSpecs.compose) {
-          component.domains = ['themok.io'];
-        }
-      }
-      const domains = getUnifiedDomainsForApp(appSpecs);
-      const customDomains = getCustomDomainsForApp(appSpecs);
-      if (appSpecs.version <= 3) {
-        const { ports } = appSpecs;
-        if (domains.length === ports.length + 1) {
-          // eslint-disable-next-line no-await-in-loop
-          const domainOperationsSuccessful = await doDomainCertOperationsForFDMdomains(domains);
-          if (domainOperationsSuccessful) {
-            log.info(`Application domain and ssl for ${appSpecs.name} is ready`);
-            appsOK.push(appSpecs);
-          } else {
-            log.error(`Domain/ssl issues for ${appSpecs.name}`);
-          }
-          if (domainOperationsSuccessful && customDomains.length) {
-            // eslint-disable-next-line no-await-in-loop
-            const customCertOperationsSuccessful = await doCertOperationsForCustomDomains(customDomains);
-            if (customCertOperationsSuccessful) {
-              log.info(`Application domain and ssl for custom domains of ${appSpecs.name} is ready`);
-            } else {
-              log.error(`Domain/ssl issues for custom domains of ${appSpecs.name}`);
-            }
-          }
-        } else {
-          log.error(`Application ${appSpecs.name} has wierd domain, settings. This is a bug.`);
-        }
-      } else {
-        // composed app
-        let ports = 0;
-        appSpecs.compose.forEach((component) => {
-          // ports += 1; // component name itself not required in new scheme
-          ports += component.ports.length;
-        });
-        if (domains.length === ports + 1) { // + 1 for app name
-          // eslint-disable-next-line no-await-in-loop
-          const domainOperationsSuccessful = await doDomainCertOperationsForFDMdomains(domains);
-          if (domainOperationsSuccessful) {
-            log.info(`Application domain and ssl for ${appSpecs.name} is ready`);
-            appsOK.push(appSpecs);
-          } else {
-            log.error(`Domain/ssl issues for ${appSpecs.name}`);
-          }
-          if (domainOperationsSuccessful && customDomains.length) {
-            // eslint-disable-next-line no-await-in-loop
-            const customCertOperationsSuccessful = await doCertOperationsForCustomDomains(customDomains);
-            if (customCertOperationsSuccessful) {
-              log.info(`Application domain and ssl for custom domains of ${appSpecs.name} is ready`);
-            } else {
-              log.error(`Domain/ssl issues for custom domains of ${appSpecs.name}`);
-            }
-          }
-        } else {
-          log.error(`Application ${appSpecs.name} has wierd domain, settings. This is a bug.`);
-        }
-      }
-    }
+    const appsOK = await processApplications(applicationSpecifications, myFDMnameORip);
     // check appsOK against mandatoryApps
     for (const mandatoryApp of mandatoryApps) {
       const appExists = appsOK.find((app) => app.name === mandatoryApp);
@@ -800,7 +254,7 @@ async function generateAndReplaceMainApplicationHaproxyConfig() {
         if (mandatoryApps.includes(app.name) && appIps.length < 1) {
           throw new Error(`Application ${app.name} checks not ok. PANIC.`);
         }
-        const domains = getUnifiedDomainsForApp(app);
+        const domains = getUnifiedDomains(app);
         const customConfigs = getCustomConfigs(app);
         if (app.version <= 3) {
           for (let i = 0; i < app.ports.length; i += 1) {
@@ -1108,10 +562,6 @@ async function initializeServices() {
 
 async function start() {
   try {
-    db = await serviceHelper.connectMongoDb();
-    const database = db.db(config.database.mainDomain.database);
-    database.collection(recordsCollection).createIndex({ ip: 1 }, { name: 'query for getting list of Flux node data associated to IP address' });
-    database.collection(recordsCollection).createIndex({ domain: 1 }, { name: 'query for getting list of Flux node data associated to Domain' });
     log.info('Initiating FDM API services...');
     initializeServices();
   } catch (e) {
@@ -1125,8 +575,4 @@ async function start() {
 
 module.exports = {
   start,
-  getAllRecordsDBAPI,
-  listDNSRecordsAPI,
-  dnsLookup,
-  isDomainPointedToThisFDM,
 };
